requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
absl-py==2.1.0
    # via mediapipe
accelerate==1.2.1
    # via controlnet-v1-1 (pyproject.toml)
aiofiles==23.2.1
    # via gradio
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   gradio
    #   httpx
    #   starlette
attrs==24.3.0
    # via mediapipe
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via sounddevice
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   typer
    #   uvicorn
contourpy==1.3.1
    # via matplotlib
controlnet-aux==0.0.9
    # via controlnet-v1-1 (pyproject.toml)
cycler==0.12.1
    # via matplotlib
diffusers==0.32.1
    # via controlnet-v1-1 (pyproject.toml)
einops==0.8.0
    # via
    #   controlnet-v1-1 (pyproject.toml)
    #   controlnet-aux
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.6
    # via gradio
ffmpy==0.5.0
    # via gradio
filelock==3.16.1
    # via
    #   controlnet-aux
    #   diffusers
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
flatbuffers==24.12.23
    # via mediapipe
fonttools==4.55.3
    # via matplotlib
fsspec==2024.12.0
    # via
    #   gradio-client
    #   huggingface-hub
    #   torch
gradio==5.12.0
    # via
    #   controlnet-v1-1 (pyproject.toml)
    #   spaces
gradio-client==1.5.4
    # via gradio
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hf-transfer==0.1.9
    # via controlnet-v1-1 (pyproject.toml)
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   gradio
    #   gradio-client
    #   safehttpx
    #   spaces
huggingface-hub==0.27.1
    # via
    #   accelerate
    #   controlnet-aux
    #   diffusers
    #   gradio
    #   gradio-client
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
imageio==2.36.1
    # via scikit-image
importlib-metadata==8.5.0
    # via
    #   controlnet-aux
    #   diffusers
jax==0.4.38
    # via mediapipe
jaxlib==0.4.38
    # via
    #   jax
    #   mediapipe
jinja2==3.1.5
    # via
    #   gradio
    #   torch
kiwisolver==1.4.8
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   gradio
    #   jinja2
matplotlib==3.10.0
    # via mediapipe
mdurl==0.1.2
    # via markdown-it-py
mediapipe==0.10.20
    # via controlnet-v1-1 (pyproject.toml)
ml-dtypes==0.5.1
    # via
    #   jax
    #   jaxlib
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via
    #   scikit-image
    #   torch
numpy==1.26.4
    # via
    #   accelerate
    #   contourpy
    #   controlnet-aux
    #   diffusers
    #   gradio
    #   imageio
    #   jax
    #   jaxlib
    #   matplotlib
    #   mediapipe
    #   ml-dtypes
    #   opencv-contrib-python
    #   opencv-python-headless
    #   pandas
    #   scikit-image
    #   scipy
    #   tifffile
    #   torchvision
    #   transformers
    #   xformers
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
opencv-contrib-python==4.10.0.84
    # via mediapipe
opencv-python-headless==4.10.0.84
    # via
    #   controlnet-v1-1 (pyproject.toml)
    #   controlnet-aux
opt-einsum==3.4.0
    # via jax
orjson==3.10.14
    # via gradio
packaging==24.2
    # via
    #   accelerate
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   lazy-loader
    #   matplotlib
    #   scikit-image
    #   spaces
    #   transformers
pandas==2.2.3
    # via gradio
pillow==11.1.0
    # via
    #   controlnet-aux
    #   diffusers
    #   gradio
    #   imageio
    #   matplotlib
    #   scikit-image
    #   torchvision
protobuf==4.25.5
    # via mediapipe
psutil==5.9.8
    # via
    #   accelerate
    #   spaces
pycparser==2.22
    # via cffi
pydantic==2.10.5
    # via
    #   fastapi
    #   gradio
    #   spaces
pydantic-core==2.27.2
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.19.1
    # via rich
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-multipart==0.0.20
    # via gradio
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   gradio
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via
    #   diffusers
    #   transformers
requests==2.32.3
    # via
    #   diffusers
    #   huggingface-hub
    #   spaces
    #   transformers
rich==13.9.4
    # via typer
ruff==0.9.1
    # via gradio
safehttpx==0.1.6
    # via gradio
safetensors==0.5.2
    # via
    #   controlnet-v1-1 (pyproject.toml)
    #   accelerate
    #   diffusers
    #   transformers
scikit-image==0.25.0
    # via controlnet-aux
scipy==1.15.1
    # via
    #   controlnet-aux
    #   jax
    #   jaxlib
    #   scikit-image
semantic-version==2.10.0
    # via gradio
sentencepiece==0.2.0
    # via mediapipe
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sounddevice==0.5.1
    # via mediapipe
spaces==0.32.0
    # via controlnet-v1-1 (pyproject.toml)
starlette==0.41.3
    # via
    #   fastapi
    #   gradio
sympy==1.13.3
    # via torch
tifffile==2025.1.10
    # via scikit-image
timm==0.6.7
    # via controlnet-aux
tokenizers==0.21.0
    # via transformers
tomlkit==0.13.2
    # via gradio
torch==2.4.0
    # via
    #   controlnet-v1-1 (pyproject.toml)
    #   accelerate
    #   controlnet-aux
    #   timm
    #   torchvision
    #   xformers
torchvision==0.19.0
    # via
    #   controlnet-v1-1 (pyproject.toml)
    #   controlnet-aux
    #   timm
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.48.0
    # via controlnet-v1-1 (pyproject.toml)
triton==3.0.0
    # via torch
typer==0.15.1
    # via gradio
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   rich
    #   spaces
    #   torch
    #   typer
    #   uvicorn
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via gradio
websockets==14.1
    # via gradio-client
xformers==0.0.27.post2
    # via controlnet-v1-1 (pyproject.toml)
zipp==3.21.0
    # via importlib-metadata
